requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
carball==0.14.7
anthropic==0.25.0
